+++ heylogs-api/src/main/java/internal/heylogs/spi/VersioningSupport.java
@@ -1,7 +1,6 @@
 package internal.heylogs.spi;
 
 import lombok.NonNull;
-import nbbrd.heylogs.Config;
 import nbbrd.heylogs.Version;
 import nbbrd.heylogs.spi.Versioning;
 
@@ -15,9 +14,12 @@ public final class VersioningSupport {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
     }
 
-    public static Stream<Versioning> versioningStreamOf(List<Versioning> list, List<Version> releases, Config config) {
-        return list.stream()
-                .filter(versioning -> releases.stream().allMatch(release -> versioning.isValidVersion(release.getRef(), config)));
+    public static Stream<Versioning> versioningStreamOf(List<Versioning> list, List<Version> releases, String versioningArg) {
+        return list.stream().filter(versioning ->
+                releases.stream()
+                        .map(Version::getRef)
+                        .allMatch(versioning.getVersioningPredicate(versioningArg))
+        );
     }
 
     public static @NonNull Predicate<Versioning> onVersioningId(@NonNull String id) {