+++ heylogs-api/src/test/java/internal/heylogs/base/ExtendedRulesTest.java
@@ -1,7 +1,9 @@
 package internal.heylogs.base;
 
+import com.vladsch.flexmark.ast.Heading;
 import com.vladsch.flexmark.ast.LinkNodeBase;
 import com.vladsch.flexmark.util.ast.Node;
+import nbbrd.design.MightBePromoted;
 import nbbrd.heylogs.Config;
 import nbbrd.heylogs.Nodes;
 import nbbrd.heylogs.spi.RuleContext;
@@ -9,8 +11,12 @@ import nbbrd.heylogs.spi.RuleIssue;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import tests.heylogs.api.Sample;
+import tests.heylogs.spi.MockedVersioning;
 
 import java.util.Objects;
+import java.util.function.Predicate;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 
 import static internal.heylogs.base.ExtendedRules.*;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -103,5 +109,48 @@ public class ExtendedRulesTest {
         assertThat(hasImbalancedBraces("[(])")).isTrue();
         assertThat(hasImbalancedBraces("{(})")).isTrue();
         assertThat(hasImbalancedBraces("(()")).isTrue();
+    }
+
+    @Test
+    public void testValidateVersioningFormat() {
+        RuleContext withoutSemver = RuleContext.DEFAULT;
+        RuleContext withSemver = RuleContext
+                .builder()
+                .config(Config
+                        .builder()
+                        .versioningId(REGEX_VERSIONING.getVersioningId())
+                        .versioningArg(Pattern.compile("^\\d+\\.\\d+\\.\\d+$").pattern())
+                        .build())
+                .versioning(REGEX_VERSIONING)
+                .build();
+
+        assertThat(Nodes.of(Heading.class).descendants(using("/InvalidSemver.md")))
+                .map(node -> VERSIONING_FORMAT.getRuleIssueOrNull(node, withoutSemver))
+                .filteredOn(Objects::nonNull)
+                .isEmpty();
+
+        assertThat(Nodes.of(Heading.class).descendants(using("/InvalidSemver.md")))
+                .map(node -> VERSIONING_FORMAT.getRuleIssueOrNull(node, withSemver))
+                .filteredOn(Objects::nonNull)
+                .hasSize(1)
+                .contains(RuleIssue.builder().message("Invalid regex format: '.1.0'").line(4).column(1).build());
+    }
+
+    @MightBePromoted
+    private static final MockedVersioning REGEX_VERSIONING = MockedVersioning
+            .builder()
+            .versioningId("regex")
+            .validation(ExtendedRulesTest::getRegexPredicate)
+            .build();
+
+    private static Predicate<CharSequence> getRegexPredicate(String regex) {
+        if (regex != null) {
+            try {
+                Pattern pattern = Pattern.compile(regex);
+                return text -> pattern.matcher(text).matches();
+            } catch (PatternSyntaxException ignore) {
+            }
+        }
+        return MockedVersioning.ignoreVersioning();
     }
 }