+++ heylogs-api/src/test/java/tests/heylogs/spi/VersioningAssert.java
@@ -1,7 +1,6 @@
 package tests.heylogs.spi;
 
 import lombok.NonNull;
-import nbbrd.heylogs.Config;
 import nbbrd.heylogs.spi.Versioning;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -23,9 +22,9 @@ public final class VersioningAssert {
                 .isNotNull();
 
         assertThatNullPointerException()
-                .isThrownBy(() -> x.isValidVersion(null, Config.DEFAULT));
+                .isThrownBy(() -> x.getVersioningPredicate(null).test(null));
 
         assertThatNullPointerException()
-                .isThrownBy(() -> x.isValidVersion("null", null));
+                .isThrownBy(() -> x.getVersioningPredicate("").test(null));
     }
 }